Pass syscall times test

+++ os/src/task/task.rs
@@ -12,7 +12,7 @@ pub struct TaskControlBlock {
     /// The task context
     pub task_cx: TaskContext,
     /// The syscall times
-    pub syscall_times: [usize; MAX_SYSCALL_NUM],
+    pub syscall_times: [u32; MAX_SYSCALL_NUM],
     /// The start time when first call
     pub start_time: usize,
 }
